requirements.txt
Flask==2.0.2
Jinja2==3.1.2
opencv_python==4.5.4.60
Pillow==9.1.1
pyzbar==0.1.9
sanic==22.3.2
flask-cors==3.0.10
numpy==1.23.1
rpi-gpio==0.7.1
requests ==2.28.1